Remove example.py from spending.git
This was an example of how to use the SR spending library.  This
library used to be in this repository, but is now in tools.git.  It
therefore no longer makes sense to have the example of how to use this
library in this repository either (there are a bunch of tools that
provide such an example in tools.git as well now).