Update cost of mech kit cable ties.
In #2006 I commented that the actual cost to me was £119.69, rather than
the previously quoted amount, because the initial estimate didn't take
VAT into account.

It's unclear whether I actually received the difference in the cheque I
received; but either way the expenditure to SR is going to end up being
the higher amount.