Encode the battery-purchase costs for sr2014
This was a massively complicated order, largely due to hobbyking, and as
a result the prices are difficult explain. See trac ticket 1880 for how
this came about.

I've taken the final cost to SR and split it between the four things
purchased in that order. Some budget massaging is required.