Trim some excess spending records.
Some complexity occurred when budget reclaim happened at the end of
SR2013 (because everyone was on fire). It looks like scarzy entered the
amount that people were _claiming_, but then paid them them amount
clamped to £75. Which was correct, but this means that spending.git
over-records how much was paid, and as result some of our money becomes
dead.

All of these changes are straightforwards except for #1665, cheque
100171, where some money was paid to me. Long story short, I settled
some debts with Lilafisch by absorbing the German team budget. However,
I'm now in a position of fiddling with the record of how much I was
paid, which has the potential to look dodgy.

This patch reduces the record of payments to me by £165.20. The new
total of cheque 100171 from these files:

./2013/06/10/Mailing-invitation-letters-to-German-teams-via-airmail-etc.yaml:cost: 16.34
./2013/06/10/jmorse-travel.yaml:cost: 15.00
./2013/06/10/linode.yaml:cost: 320.00
./2013/06/10/Shipping-to-EMM-etc.yaml:cost: 190.06
./2013/06/10/Replacment-power-board-for-SEN-etc.yaml:cost: 21.45
./2013/06/10/1-sheet-of-4x2-6mm-ply.yaml:cost: 13.18
./2013/06/10/Arduino-and-Sonar-etc.yaml:cost: 75
./2013/06/10/tape-purchase-and-delivery.yaml:cost: 222.44
./2013/06/10/lilafisch-travel-2013.yaml:cost: 147.76

Totals £1021.23. This is the actual amount paid in cheque 100171, I can
show the stub to anyone willing to see. Thus, IMO, this means the
adjustment only corrects our record of what was paid, it doesn't hide
extra money that was paid to me.