Record the final state of SR2015 board comp orders
Some high speed substitutions of supplier had to be made while placing
the SR2015 component orders. Unfortunately this resulted in a slightly
higher spend (and a spend with Digikey, who we weren't previously going
to use.) I've added the receipts and amounts for what was actually spent
and where but left all of the transactions that don't match the tickets
as pending. The next step is to review the differences and submit any
necessary budget mods.